requirements_old.txt
# RestoreDoc Flask Server Requirements

# Core Flask dependencies
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# OpenAI API client
openai==1.6.1

# Environment variables
python-dotenv==1.0.0

# Additional utilities
requests==2.31.0
Pillow==10.1.0  # For image processing
reportlab==4.1.0  # For PDF generation
python-dateutil==2.8.2

# Supabase client
supabase==2.0.0

# Production server (required for Render deployment)
gunicorn==21.2.0

# Testing (optional for production)
pytest==7.4.3
pytest-flask==1.3.0